+++ tests/test_decoder.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 """This module has tests for the pvl decoder functions."""
 
-# Copyright 2019, Ross A. Beyer (rbeyer@seti.org)
+# Copyright 2019-2021, Ross A. Beyer (rbeyer@seti.org)
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.